Static `DBOS` (#122)
Use a global DBOS, which opens up the registration order flexibility
that we have in TS and PY. (Adding multiple instances would not be that
hard, we would pass the DBOS.Instance object around only in
initialization and initiation code, everything else would use the
context).

Also:
- improve sleep (call sites and use `Duration`)

Fixes #118 
Fixes #99 
Fixes #98

+++ src/test/java/dev/dbos/transact/execution/ExecutingService.java
@@ -10,7 +10,7 @@ public interface ExecutingService {
 
   String stepTwo(String input);
 
-  void sleepingWorkflow(float seconds);
+  void sleepingWorkflow(double seconds);
 
   void setExecutingService(ExecutingService service);
